WIP Fix mandatarissen and personen selection
This commit fixes the selection of mandatarissen or persons and meeting
attendees. We can now select them and insert them into a document.
Still need to fix the mandatarissen display in the pulgin, being able to
modify the list once inserted and being able to choose a president and a
secretary in the mandatarissen list.