Update data for card when card is rendered
It seems the rendering of the card triggers didReceiveAttrs, so we
fetch the bestuurseenheid at that point in time.  This ensures we
fetch the data at the right time.

Note that there is still duplication going on in that the service is
also trying to find the same data.  Difference is that it
is (incorrectly) trying to do so on first highlight instead.