Only show "Beheer aanwezigen" button when data is available
Wher there's no bestuurseenheidUri, we don't render the button
anymore.  It would only make the plugin crash and the hint provides
sufficient information.  This approach isn't perfect in that we assume
we'll always find the right info when a URI is found in the database.
This would be an easy enhancement.