Prevent infinite rendering loop
The aanwezigenToSelect depended on itself which caused an infinite
loop.  This has been altered so it's now based on the properties it
uses itself.  There's a strange thing going on with an intermediate
"_aanwezigen" which seems to be allowed but not sure if this should
actually be a dependent key.  We're happy that it works now as that
was the main goal.